% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_coda_tab.R
\name{create_coda_tab}
\alias{create_coda_tab}
\title{Build the "Compositional Analysis" tab's UI}
\usage{
create_coda_tab()
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Compositional Analysis" tab's UI
}
