% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_evs_tab.R
\name{create_evs_tab}
\alias{create_evs_tab}
\title{Build the "Extreme Value Analysis" tab's UI}
\usage{
create_evs_tab()
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Extreme Value Analysis" tab's UI
}
