% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{create_histogram_density}
\alias{create_histogram_density}
\title{Create a histogram with an overlaid density curve}
\usage{
create_histogram_density(
  data,
  column,
  bins = 30,
  title = NULL,
  xlab = NULL,
  ylab = "Density"
)
}
\arguments{
\item{data}{A data frame.}

\item{column}{Name of the numeric column to plot.}

\item{bins}{Number of histogram bins. Default 30.}

\item{title}{Plot title. Defaults to \code{"Distribution of <column>"}.}

\item{xlab}{X-axis label. Defaults to \code{column}.}

\item{ylab}{Y-axis label. Default \code{"Density"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a histogram with an overlaid density curve
}
