% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inclusion_presets.R
\name{create_inclusion_presets_panel}
\alias{create_inclusion_presets_panel}
\title{Build the controls and outputs of the "Inclusion Presets" tab}
\usage{
create_inclusion_presets_panel(ns)
}
\arguments{
\item{ns}{The tab's namespace function (\code{NS(id)}).}
}
\value{
A \code{shiny::fluidRow()} with the controls (left) and the diagram,
tables and downloads (right).
}
\description{
Build the controls and outputs of the "Inclusion Presets" tab
}
