% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inclusion_tab.R
\name{create_inclusion_tab}
\alias{create_inclusion_tab}
\title{Build the "Inclusion Presets" tab's UI}
\usage{
create_inclusion_tab(id)
}
\arguments{
\item{id}{Module namespace id - must match the id passed to
\code{moduleServer()} for this tab in \code{server_logic.R}.}
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Inclusion Presets" tab's UI
}
