% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_point_pattern_analysis.R
\name{create_kernel_intensity_plot}
\alias{create_kernel_intensity_plot}
\title{Plot the kernel intensity ("hotspot") map}
\usage{
create_kernel_intensity_plot(intensity_result)
}
\arguments{
\item{intensity_result}{A result from \code{compute_kernel_intensity()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the kernel intensity ("hotspot") map
}
