% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_point_pattern_analysis.R
\name{create_l_function_plot}
\alias{create_l_function_plot}
\title{Plot the L-function as L(r) - r (the conventional clustering diagnostic)}
\usage{
create_l_function_plot(l_result)
}
\arguments{
\item{l_result}{A result from \code{compute_l_function()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the L-function as L(r) - r (the conventional clustering diagnostic)
}
