% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_multiple_ternary_tab.R
\name{create_multiple_ternary_tab}
\alias{create_multiple_ternary_tab}
\title{Build the "Multiple Ternary Creator" tab's UI}
\usage{
create_multiple_ternary_tab(id)
}
\arguments{
\item{id}{Module namespace id - must match the id passed to
\code{moduleServer()} for this tab in \code{server_logic.R}.}
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Multiple Ternary Creator" tab's UI
}
