% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_spatial_tab.R
\name{create_spatial_tab}
\alias{create_spatial_tab}
\title{Build the "Spatial Clustering" tab's UI}
\usage{
create_spatial_tab()
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Spatial Clustering" tab's UI
}
