% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_ternary_plots_tab.R
\name{create_ternary_plots_tab}
\alias{create_ternary_plots_tab}
\title{Build the "Ternary Plots" tab's UI}
\usage{
create_ternary_plots_tab(id)
}
\arguments{
\item{id}{Module namespace id - must match the id passed to
\code{moduleServer()} for this tab in \code{server_logic.R}.}
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Ternary Plots" tab's UI
}
