% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_performance_summary}
\alias{get_performance_summary}
\title{Report the duration of the most recent \code{start_performance_monitor()} run}
\usage{
get_performance_summary()
}
\value{
A formatted character string, or a "not available" message if
no monitor has been started.
}
\description{
Report the duration of the most recent \code{start_performance_monitor()} run
}
