% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_input_helpers.R
\name{inclusion_detected_elements}
\alias{inclusion_detected_elements}
\title{Element symbols with a wt\% column in a data frame}
\usage{
inclusion_detected_elements(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector of element symbols (order of the columns).
}
\description{
Element symbols with a wt\% column in a data frame
}
