% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{MIN_POINT_SIZE}
\alias{MIN_POINT_SIZE}
\title{App-wide options and constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_POINT_SIZE
}
\description{
Point-size bounds shared by every plot's point-size legend/scaling
(ternary_plot_save.R, ternary_plot_preview.R).
}
\keyword{datasets}
