% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_progress}
\alias{update_progress}
\title{Report progress within the current step}
\usage{
update_progress(step_number, message)
}
\arguments{
\item{step_number}{Current step number.}

\item{message}{Progress message to print.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Report progress within the current step
}
