% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vidternary-package.R
\docType{package}
\name{vidternary-package}
\alias{vidternary}
\alias{vidternary-package}
\title{vidternary: Modular Ternary Plot Analysis Tool}
\description{
An interactive Shiny application and supporting function library for
compositional / ternary-diagram analysis of nonmetallic-inclusion
chemistry from automated SEM/EDS data: ternary and hexagonal-ternary
plotting, log-ratio (CLR / ILR) compositional analysis, Mahalanobis and
Isolation-Forest outlier detection, IQR / Z-score / MAD filtering,
extreme-value (Murakami / ASTM E2283) inclusion rating, and spatial
point-pattern analysis (Clark-Evans, Ripley's K/L/G, CSR envelope,
kernel intensity).
}
\author{
\strong{Maintainer}: Vid Kuder Marušič \email{vid.kudermarusic@gmail.com}

}
\keyword{internal}
