% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_hex_ternary_tab.R
\name{create_hex_ternary_tab}
\alias{create_hex_ternary_tab}
\title{Build the "Hexagonal Ternary Diagram" tab's UI}
\usage{
create_hex_ternary_tab()
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Hexagonal Ternary Diagram" tab's UI
}
