% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_plot_builder_tab.R
\name{create_plot_builder_tab}
\alias{create_plot_builder_tab}
\title{Build the "Plot Builder" tab's UI}
\usage{
create_plot_builder_tab()
}
\value{
A \code{shiny::tabPanel()}.
}
\description{
Build the "Plot Builder" tab's UI
}
