% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{MIN_POINT_SIZE}
\alias{MIN_POINT_SIZE}
\title{App-wide options and constants}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_POINT_SIZE
}
\description{
This file contains all the constants and configuration values used throughout the application.
Centralizing these makes it easier to maintain and modify the app behavior.
}
\keyword{datasets}
